+++ actionpack/lib/action_dispatch/routing.rb
@@ -1,5 +1,5 @@
 # encoding: UTF-8
-require 'active_support/core_ext/object/to_param'
+require 'active_support/core_ext/object/to_query'
 require 'active_support/core_ext/regexp'
 require 'active_support/dependencies/autoload'
