+++ activesupport/lib/active_support/core_ext/array/conversions.rb
@@ -1,7 +1,6 @@
 require 'active_support/xml_mini'
 require 'active_support/core_ext/hash/keys'
 require 'active_support/core_ext/string/inflections'
-require 'active_support/core_ext/object/to_param'
 require 'active_support/core_ext/object/to_query'
 
 class Array
