+++ activesupport/lib/active_support/core_ext/hash/conversions.rb
@@ -1,7 +1,6 @@
 require 'active_support/xml_mini'
 require 'active_support/time'
 require 'active_support/core_ext/object/blank'
-require 'active_support/core_ext/object/to_param'
 require 'active_support/core_ext/object/to_query'
 require 'active_support/core_ext/array/wrap'
 require 'active_support/core_ext/hash/reverse_merge'
