+++ activesupport/lib/active_support/core_ext/object.rb
@@ -9,6 +9,5 @@ require 'active_support/core_ext/object/conversions'
 require 'active_support/core_ext/object/instance_variables'
 
 require 'active_support/core_ext/object/json'
-require 'active_support/core_ext/object/to_param'
 require 'active_support/core_ext/object/to_query'
 require 'active_support/core_ext/object/with_options'
