+++ activesupport/lib/active_support/core_ext/object/conversions.rb
@@ -1,4 +1,3 @@
-require 'active_support/core_ext/object/to_param'
 require 'active_support/core_ext/object/to_query'
 require 'active_support/core_ext/array/conversions'
 require 'active_support/core_ext/hash/conversions'
