--- activesupport/lib/active_support/core_ext/object/to_param.rb
@@ -1,1 +1,0 @@
-require 'active_support/core_ext/object/to_query'
